+++ NOTICE.libmdbx.md
@@ -1,0 +1,39 @@
+libmdbx (aka MDBX) is an extremely fast, compact, powerful, embeddedable,
+transactional key-value storage engine with open-source code. MDBX has a
+specific set of properties and capabilities, focused on creating unique
+lightweight solutions.
+
+Please visit https://libmdbx.dqdkfa.ru for more information, changelog,
+documentation, C++ API description and links to the original git repo
+with the source code. Questions, feedback and suggestions are welcome
+to the Telegram' group https://t.me/libmdbx.
+
+Donations are welcome to the Ethereum/ERC-20 `0xD104d8f8B2dC312aaD74899F83EBf3EEBDC1EA3A`.
+Всё будет хорошо!
+
+Copyright 2015-2025 Леонид Юрьев aka Leonid Yuriev <leo@yuriev.ru>
+SPDX-License-Identifier: Apache-2.0
+For notes about the license change, credits and acknowledgments,
+please refer to the COPYRIGHT file within libmdbx source.
+
+---
+
+On 2022-04-15, without any warnings or following explanations, the
+Github administration deleted _libmdbx_, my account and all other
+projects (status 404). A few months later, without any involvement or
+notification from/to me, the projects were restored/opened in the "public
+read-only archive" status from some kind of incomplete backup. I regard
+these actions of Github as malicious sabotage, and I consider the Github
+service itself to have lost trust forever.
+
+As a result of what has happened, I will never, under any circumstances,
+post the primary sources (aka origins) of my projects on Github, or rely
+in any way on the Github infrastructure.
+
+Nevertheless, realizing that it is more convenient for users of
+_libmdbx_ and other my projects to access ones on Github, I do not want
+to restrict their freedom or create inconvenience, and therefore I place
+mirrors (aka mirrors) of such repositories on Github since 2025. At the
+same time, I would like to emphasize once again that these are only
+mirrors that can be frozen, blocked or deleted at any time, as was the
+case in 2022.
