+++ libmdbx/.clang-format-ignore
