+++ libmdbx/ChangeLog.md
@@ -4,6 +4,45 @@ ChangeLog
 English version [by liar Google](https://libmdbx-dqdkfa-ru.translate.goog/md__change_log.html?_x_tr_sl=ru&_x_tr_tl=en)
 and [by Yandex](https://translated.turbopages.org/proxy_u/ru-en.en/https/libmdbx.dqdkfa.ru/md__change_log.html).
 
+## v0.13.6 в процессе накопления изменений
+
+Поддерживающий выпуск стабильной ветки с исправлением обнаруженных ошибок и устранением недочётов.
+
+Благодарности:
+
+ - [Erigon](https://erigon.tech/) за спонсорство.
+ - [Илье Михееву](https://t.me/IlyaMkhv) и команде [Erigon](https://github.com/erigontech) за сообщения о проблеме и тестирование.
+
+Исправления:
+
+ - Устраненен регресс при использовании курсоров для DBI=0 в читающих транзакциях.
+
+   В результате рефакторинга и ряда оптимизаций для завершения/гашения
+   курсоров в читающих и пишущих транзакций стал использоваться общий код.
+   Причем за основу, был взят соответствующий фрагмент относящийся к
+   пишущим транзакциям, в которых пользователю не позволяется
+   использоваться курсоры для DBI=0 и поэтому эта итераций пропускалась.
+
+   В результате, при завершении читающих транзакциях, курсоры связанные с
+   DBI=0 не завершались должным образом, а при их повторном использовании
+   или явном закрытии после завершения читающей транзакции происходило
+   обращение к уже освобожденной памяти. Если же такие курсоры
+   отсоединялись или закрывались до завершения читающей транзакции, то
+   ошибка не имела шансов на проявление.
+
+Прочие доработки:
+
+ - Уточнен тип адреса для пожертвований.
+   Ethereum/ERC-20 позволяет перечислять не только ETH, но и другие валюты/токены, в том числе USDC.
+
+ - Вброс `std::invalid_argument` теперь производится явным сообщением `MDBX_EINVAL`.
+
+ - Дополнен тест курсоров extra/cursor-closing.
+
+ - В `NOTICE` обновлена информация о Github.
+
+--------------------------------------------------------------------------------
+
 ## v0.13.5 "Труба" от 2025-03-21
 
 Поддерживающий выпуск стабильной ветки с исправлением обнаруженных ошибок и устранением недочётов.
@@ -107,7 +146,7 @@ and [by Yandex](https://translated.turbopages.org/proxy_u/ru-en.en/https/libmdbx
    поломать родительскую, сделав её продолжение невозможным. Если восстанавливать, то также следует «воскрешать» закрытые
    курсоры, что неизбежно приведет к путанице, утечкам памяти и использованию после освобождения.
 
- - В C++ API отменён вброс исключения при запросе транзакции у отсоединённого курсора посредством вывоза `mdbx::cursor::txn()`.
+ - В C++ API отменён вброс исключения при запросе транзакции у отсоединённого курсора посредством вызова `mdbx::cursor::txn()`.
 
 Прочие доработки:
