+++ libmdbx/LICENSE
