+++ libmdbx/VERSION.json
@@ -1,0 +1,1 @@
+{ "git_describe": "v0.13.5-6-g19dc93fc", "git_timestamp": "2025-03-23T17:46:28+03:00", "git_tree": "85af2a025731d4f83e6cc576b1b335db54992781", "git_commit": "19dc93fc76590dd309660fc10802b1f9e3f71f4c", "semver": "0.13.5.6" }
