+++ libmdbx/keep.go
@@ -1,0 +1,3 @@
+package libmdbx
+
+import "C"
