+++ libmdbx/man1/mdbx_stat.1
