--- mdbxdist/.gitignore
@@ -1,47 +1,0 @@
-*.[ao]
-*.bak
-*.exe
-*.err
-*.gcda
-*.gcno
-*.gcov
-*.lo
-*.orig
-*.rej
-*.so
-*.dll
-*.dylib
-*.dSYM
-*[~#]
-.idea
-.vs/
-.vscode/
-cmake-build-*
-@*
-core
-mdbx_example
-libmdbx.creator.user
-CMakeLists.txt.user
-mdbx_chk
-mdbx_copy
-mdbx_drop
-mdbx_dump
-mdbx_load
-mdbx_stat
-mdbx_test
-test.log
-test/tmp.db
-test/tmp.db-lck
-tmp.db
-tmp.db-lck
-valgrind.*
-src/version.c
-src/config.h
-dist/
-dist-check/
-dist-checked.tag
-*.tar*
-*.zip
-docs/Doxyfile
-docs/html/
-buildflags.tag
