--- mdbxdist/VERSION.json
@@ -1,1 +1,0 @@
-{ "git_describe": "v0.13.5-0-ge3324cef", "git_timestamp": "2025-03-21T21:14:00+03:00", "git_tree": "eaa12f8b818d12a92de2b9a3bde3b5ea223953a0", "git_commit": "e3324cef918407fe48c90e513577ec99a0d15b62", "semver": "0.13.5" }
