--- mdbxdist/VERSION.txt
@@ -1,1 +1,0 @@
-0.12.13.0
