--- mdbxdist/keep.go
@@ -1,3 +1,0 @@
-package mdbxdist
-
-import "C"
